+++ plugin/src/net/groboclown/idea/p4ic/server/connection/ClientPasswordConnectionHandler.java
@@ -13,25 +13,18 @@
  */
 package net.groboclown.idea.p4ic.server.connection;
 
-import com.intellij.ide.passwordSafe.PasswordSafe;
-import com.intellij.ide.passwordSafe.PasswordSafeException;
-import com.intellij.ide.passwordSafe.ui.PasswordSafePromptDialog;
-import com.intellij.openapi.application.ModalityState;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
 import com.perforce.p4java.PropertyDefs;
 import com.perforce.p4java.exception.AccessException;
-import com.perforce.p4java.exception.ConnectionException;
 import com.perforce.p4java.exception.P4JavaException;
 import com.perforce.p4java.option.server.LoginOptions;
 import com.perforce.p4java.server.IOptionsServer;
-import net.groboclown.idea.p4ic.P4Bundle;
 import net.groboclown.idea.p4ic.config.ServerConfig;
-import net.groboclown.idea.p4ic.extension.P4Vcs;
 import net.groboclown.idea.p4ic.server.ConfigurationProblem;
 import net.groboclown.idea.p4ic.server.ConnectionHandler;
 import net.groboclown.idea.p4ic.v2.server.connection.AlertManager;
-import net.groboclown.idea.p4ic.v2.ui.alerts.PasswordRequiredHandler;
+import net.groboclown.idea.p4ic.v2.server.connection.PasswordManager;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -69,10 +62,10 @@ public class ClientPasswordConnectionHandler extends ConnectionHandler {
     @Override
     public void defaultAuthentication(@Nullable Project project, @NotNull IOptionsServer server, @NotNull ServerConfig config)
             throws P4JavaException {
-        // Default login - use the user provided password from the config
-        String password = config.getPlaintextPassword();
+        // Default login - use the simple password
+        String password = PasswordManager.getInstance().getSimplePassword(config);
 
-        if (password != null && password.length() > 0) {
+        if (password != null) {
             // If the password is blank, then there's no need for the
             // user to log in; in fact, that wil raise an error by Perforce
             try {
@@ -81,7 +74,7 @@ public class ClientPasswordConnectionHandler extends ConnectionHandler {
                 if (ex.getServerMessage().hasMessageFragment("'login' not necessary")) {
                     // ignore login and keep going
                     LOG.info(config + ": User provided password, but  it is not necessary", ex);
-                    // TODO tell the caller that the password should be forgotten
+                    // FIXME send a user alert about how they need to remove the password setting.
                 } else {
                     throw ex;
                 }
@@ -89,34 +82,21 @@ public class ClientPasswordConnectionHandler extends ConnectionHandler {
         }
     }
 
+
     @Override
     public boolean forcedAuthentication(@Nullable Project project, @NotNull IOptionsServer server,
             @NotNull ServerConfig config, @NotNull AlertManager alerts) throws P4JavaException {
-        try {
-            LOG.debug("Asking PasswordSafe for the password");
-            String password = PasswordSafe.getInstance().getPassword(project,
-                    P4Vcs.class, config.getServiceName());
-            if (password != null && password.length() > 0) {
-                // If the password is blank, then there's no need for the
-                // user to log in; in fact, that wil raise an error by Perforce
-                try {
-                    server.login(password, new LoginOptions(false, true));
-                    LOG.debug("No issue logging in with stored password");
-                } catch (AccessException e) {
-                    LOG.debug("Stored password was bad; forgetting it");
-                    PasswordSafe.getInstance().removePassword(project,
-                            P4Vcs.class, config.getServiceName());
-                    throw e;
-                }
+        String password = PasswordManager.getInstance().getPassword(project, config);
+        if (password != null) {
+            try {
+                server.login(password, new LoginOptions(false, true));
+                LOG.debug("No issue logging in with stored password");
                 return true;
-            } else {
-                LOG.debug("No password found");
-                alerts.addCriticalError(new PasswordRequiredHandler(project, config), null);
+            } catch (AccessException e) {
+                LOG.info("Stored password was bad; forgetting it");
+                PasswordManager.getInstance().forgetPassword(project, config);
+                throw e;
             }
-        } catch (PasswordSafeException e) {
-            LOG.debug("PasswordSafe access caused an error", e);
-            // FIXME have a better exception
-            throw new ConnectionException(e);
         }
         return false;
     }
@@ -127,23 +107,6 @@ public class ClientPasswordConnectionHandler extends ConnectionHandler {
         // This config only uses the fields that are required in the
         // server config.  No additional checks are needed.
         return Collections.emptyList();
-    }
-
-
-    /**
-     * Not really the best place for this method, but it centralizes the PasswordSafe
-     * access.
-     *
-     * @return true if the user entered a password
-     */
-    public static boolean askPassword(@Nullable Project project, @NotNull final ServerConfig config) {
-        return PasswordSafePromptDialog.askPassword(project, ModalityState.any(),
-                P4Bundle.message("login.password.title"),
-                P4Bundle.message("login.password.message", config.getServiceName(), config.getUsername()),
-                P4Vcs.class, config.getServiceName(),
-                true,
-                P4Bundle.message("login.password.error")
-                ) != null;
     }
 
 }